requirements.txt
setuptools==41.2.0
Pillow==6.2.1
